requirements.txt
pandas>=1.3.0
datasets>=2.10.0
google-generativeai>=0.3.2
pytest>=7.0.0